feat(cli): Add support for mTLS
Adds support for TLS configuration. This adds a TLS section to the API
configuration object. It allows specifying InsecureSkipVerify, an
additional CA cert, and client certs.
Client certs are not added as an API authentication method, since they
are often used in together with these other authentication methods.

Global CLI options have been added to overwrite the API specific options
and to be able to use them during API setup. The values used during
initial setup will be written to the config.
An additional CLI wizard/asker is also provided.

As part of this change, a small refactoring on the CLI flags handling
has been done: a secondary flagset with fixed global flags is created.
This flagset can be parsed before API discovery, without getting in the
way of cobra.