auth: support deferring authentication to a third party script
Some OpenAPI backed APIs implement a custom API signature scheme
to provide stronger guarantees against replay attacks than those
provided by OAuth.

Generally speaking, these APIs tend to either wire a computed
signature in an HTTP header or in a query argument in the URL.

This patch implements a new `external-tool` auth method which
executes a configured program against a JSON representation of
the request built by restish. If the program returns any output
it is deserialized from JSON, allowing for new headers to be set,
or for the URL to be modified. This then allows small script to
be used alongside restish to implement any sort of API signature
mechanism.